Factor out new module for test stuff, so I can use it in smoke tests
