+++ common-test/src/main/java/io/strimzi/test/k8s/KubeClusterException.java
@@ -2,7 +2,7 @@
  * Copyright 2017-2018, Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.test;
+package io.strimzi.test.k8s;
 
 public class KubeClusterException extends RuntimeException {
     public final int statusCode;
