+++ common-test/src/main/java/io/strimzi/test/k8s/Minikube.java
@@ -2,11 +2,9 @@
  * Copyright 2017-2018, Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.test;
+package io.strimzi.test.k8s;
 
-import static io.strimzi.test.Exec.exec;
-import static io.strimzi.test.Exec.execOutput;
-import static io.strimzi.test.Exec.isExecutableOnPath;
+import static io.strimzi.test.k8s.Exec.exec;
 
 /**
  * A {@link KubeCluster} implementation for {@code minikube} and {@code minishift}.
@@ -32,12 +30,12 @@ public class Minikube implements KubeCluster {
 
     @Override
     public boolean isAvailable() {
-        return isExecutableOnPath(cmd);
+        return Exec.isExecutableOnPath(cmd);
     }
 
     @Override
     public boolean isClusterUp() {
-        String output = execOutput(cmd, "status");
+        String output = Exec.execOutput(cmd, "status");
         return output.contains("minikube: Running")
                 && output.contains("cluster: Running")
                 && output.contains("kubectl: Correctly Configured:");
@@ -45,12 +43,12 @@ public class Minikube implements KubeCluster {
 
     @Override
     public void clusterUp() {
-        exec(cmd, "start");
+        Exec.exec(cmd, "start");
     }
 
     @Override
     public void clusterDown() {
-        exec(cmd, "stop");
+        Exec.exec(cmd, "stop");
     }
 
     @Override
