+++ topic-controller/pom.xml
@@ -66,6 +66,12 @@
             <version>${vertx.version}</version>
         </dependency>
         <dependency>
+            <groupId>io.strimzi</groupId>
+            <artifactId>common-test</artifactId>
+            <version>${project.version}</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
             <version>${junit.version}</version>
